Revert "msm: kgsl: Affine IRQ and worker kthread to the big CPU cluster"
This is wrong. This causes 3 IRQs to be affined to the big CPU cluster,
not just the primary kgsl_3d0_irq one. As a result, the perf crit API
thinks that the 2 extra IRQs are critical and will balance them despite
them being rarely used (kgsl_hfi_irq and kgsl_gmu_irq).

This reverts commit 60c95ae57edca9a53c69824b4000f12a64d11141.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: rany <ranygh@riseup.net>